refactor(hosting)!: value API on bare @aws-blocks/hosting; CDK on /constructs; drop /secret (review Blocking P2 #4)
Make the package's `.` entry the CDK-free value API (secret/config, getSecret/
getConfig, the CLI core, typegen) so it can be imported into an SSR/runtime bundle
without pulling in aws-cdk-lib. The HostingConstruct AND the CDK resolution engine
(partitionEnvironment/wireManagedValue/resolveSecretsAtSynth/… + their types) move
to `@aws-blocks/hosting/constructs`. The `/secret` subpath is removed — no subpath
appears in application code (key change 5).

- hosting: new constructs/index.ts (CDK barrel: construct + resolution engine +
  manifest types); index.ts trimmed to the value API; package.json repoints
  /constructs and drops /secret; secret-exports.ts deleted; typegen augments the
  bare `@aws-blocks/hosting` entry only (DEFAULT_TYPEGEN_MODULES/MARKER_MODULES).
- core/pipeline: repoint the resolution-engine + CDK-type imports to /constructs;
  value-API imports move from /secret to the bare entry.
- create-blocks-app / test-app / docs: read via `@aws-blocks/hosting`.
- Regenerated core/API.md; fixed pre-existing lint errors surfaced in changed
  files (useImportType in types.ts, node: protocol, an unused import).
- blocks vendorize-map test skips `hosting` (infra, like core).

Unpublished, so the export-map change is safe. Builds + unit suites green across
hosting/core/pipeline/blocks; the hosting-ssr test app type-checks via the bare entry.
