+++ packages/blocks/src/vendorize-map.test.ts
@@ -48,8 +48,15 @@ describe('aws-blocks.vendorize map', () => {
     for (const dep of deps) {
       // Skip infrastructure packages (not user-facing BBs). bb-lambda-compute
       // is the internal default compute — CDK-only and not customer-instantiable
-      // — so it has no public symbol to vendorize.
-      if (dep.endsWith('/core') || dep.endsWith('/auth-common') || dep.endsWith('/bb-lambda-compute')) continue;
+      // — so it has no public symbol to vendorize. `hosting` is the CDK hosting
+      // library whose value API is re-exported through core/blocks — infra, not a BB.
+      if (
+        dep.endsWith('/core') ||
+        dep.endsWith('/auth-common') ||
+        dep.endsWith('/bb-lambda-compute') ||
+        dep.endsWith('/hosting')
+      )
+        continue;
 
       // Find the package directory by resolving its main entry and walking up
       let pkgDir: string | null = null;
