+++ packages/core/src/index.ts
@@ -14,7 +14,7 @@ export {
 	type SecretValue,
 	secret,
 	type ValueKind,
-} from '@aws-blocks/hosting/secret';
+} from '@aws-blocks/hosting';
 export { type ApiHandler, ApiNamespace, type BlocksContext } from './api.js';
 export { _resetConfigCache, getConfig, getConfigSync, loadConfigToProcessEnv, preloadConfig } from './common/config.js';
 export { type BuildingBlockMeta, Scope, type ScopeOptions, type ScopeParent } from './common/index.js';
