+++ packages/core/src/pipeline.test.ts
@@ -3,9 +3,9 @@
 
 import assert from 'node:assert';
 import { afterEach, describe, it } from 'node:test';
-import { _setSynthSecretFetcher } from '@aws-blocks/hosting';
-import { config } from '@aws-blocks/hosting/secret';
-import * as cdk from 'aws-cdk-lib';
+import { _setSynthSecretFetcher } from '@aws-blocks/hosting/constructs';
+import { config } from '@aws-blocks/hosting';
+import type * as cdk from 'aws-cdk-lib';
 import { App, Stack } from 'aws-cdk-lib';
 import { Pipeline } from './pipeline/index.js';
 import type { PipelineStageConfig } from './pipeline/index.js';
