+++ packages/core/src/scripts/config.ts
@@ -11,7 +11,7 @@
  * @module
  */
 
-import { listValues, removeValue, runValueCli, setValue } from '@aws-blocks/hosting/secret';
+import { listValues, removeValue, runValueCli, setValue } from '@aws-blocks/hosting';
 import { blocksConfigPrefix } from '../secret-naming.js';
 
 /** Set (create or overwrite) a Blocks config value. */
