+++ packages/core/src/scripts/dev-server.ts
@@ -126,7 +126,7 @@ export interface DevServerOptions {
  */
 async function startTypegenWatch(): Promise<(() => void) | undefined> {
   try {
-    const { scanValueKeys, watchHostingValues } = await import('@aws-blocks/hosting/secret');
+    const { scanValueKeys, watchHostingValues } = await import('@aws-blocks/hosting');
     const scan = await scanValueKeys();
     if (scan.secretKeys.length === 0 && scan.configKeys.length === 0) {
       return undefined; // app doesn't use secret()/config() → nothing to type, skip silently
