+++ packages/hosting/src/constructs/hosting_construct.standalone.test.ts
@@ -3,9 +3,9 @@ import { afterEach, beforeEach, describe, it } from 'node:test';
 import { App, Stack } from 'aws-cdk-lib';
 import { Match, Template } from 'aws-cdk-lib/assertions';
 import { Certificate } from 'aws-cdk-lib/aws-certificatemanager';
-import * as fs from 'fs';
-import * as os from 'os';
-import * as path from 'path';
+import * as fs from 'node:fs';
+import * as os from 'node:os';
+import * as path from 'node:path';
 import type { DeployManifest } from '../manifest/types.js';
 import { config, secret } from '../secret.js';
 import { HostingConstruct } from './hosting_construct.js';
