+++ packages/hosting/src/sub_path_exports.test.ts
@@ -17,11 +17,7 @@ void describe('sub-path exports', () => {
     assert.ok(require.resolve('@aws-blocks/hosting/error'));
   });
 
-  void it('resolves ./secret', () => {
-    assert.ok(require.resolve('@aws-blocks/hosting/secret'));
-  });
-
-  void it('resolves main entry', () => {
+  void it('resolves main entry (the CDK-free value API)', () => {
     assert.ok(require.resolve('@aws-blocks/hosting'));
   });
 });
