+++ packages/pipeline/src/types.ts
@@ -1,8 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import type { KindStoreOptions } from '@aws-blocks/hosting';
-import type { ConfigValue, SecretValue } from '@aws-blocks/hosting/secret';
+import type { KindStoreOptions } from '@aws-blocks/hosting/constructs';
+import type { ConfigValue, SecretValue } from '@aws-blocks/hosting';
 import type * as cdk from 'aws-cdk-lib';
 import type * as codebuild from 'aws-cdk-lib/aws-codebuild';
 import type { ISecret } from 'aws-cdk-lib/aws-secretsmanager';
